Add helper to distort display of step bar
The intended distortion is to make each segment of the progress bar a given
minimum value so it is visible. Since a progress bar has a fixed width, we
have to decrease all the other elements of the bar by some amount in order
to accommodate the increases to elements below the minimum.