Allow prediction methods to share loaded data
It takes 2ms to output the JSON for a deploy and 450ms
with predictions. After this change it takes 250ms because
the large number of prediction rows are only loaded once.
There are other gains to be made by pushing more of the
statistics down to SQL, but this is good enough for now.